Create a watcher on the source directory
We need to create a watcher that tracks the directory that
contains the files to be transferred. If a new file is added to
the tracked directory, an event will be generated that the watcher will
notice and so the spooler will know that it has a new file to transfer.