v.1.18: Add det_composition and fix registration bug
- add det_composition
- this new metadata contains a list of detectors
- if the detector list contains a single detector, then the EPN2EOS
appends the period name with the detector and registers it in the
proper single-detector directory in the catalogue
- Example: det_composition: TPC
	   LHCPeriod: OCT
  Final period: OCT_TPC
- if there is more than one detector or the metadata det_composition
is missing, then the run is considered global and only the period
name is used.
- Example: det_composition: TPC, ITS
           LHCPeriod: OCT
  Final period: OCT

- fix registration bug
- if the transfer is successful but the registration in the catalogue
fails, the surl metadata is changed and the number of failed
registrations is appended to the surl.

Signed-off-by: Alice Suiu <alicesuiu17@gmail.com>