v.1.27: Add support for fallback storage
Add new config params:
- default SE name
- default SE io.daemons
- fallback SE name
- fallback SE io.daemons
- threshold - measured in GB; The files are default transferred to
the default SE. But, if the total size of the files in the transfer
queues is greater than the threshold, then the files will be transferred
to the fallback SE. Also, after each successful registration of a file,
it will be mirrored on the default SE (ALICE::CERN::EOSALICEO2)

Add a method that determines dynamically the storage to which the files
are transferred depending on the configured threshold.

Add two new monitoring metrics (write_Status and write_Message) that send
to MonaLisa the current active storage for transfer.

Signed-off-by: Alice Suiu <alicesuiu17@gmail.com>